requirements.txt.txt
streamlit
pandas
langchain
langchain-groq
python-dotenv